Start ElasticFEModel
Address #49

Start of the ElasticFEModel class. The class generates a square grid
using Gmsh in a similar way to FiPy.

The prototype SfePy example ("elasticity/elasticiy.py") has been
compared with OOF and broadly agrees when comparing roughly for
displacement values in similar spatial locations. The SfePy example
has the wrong boundary conditions for the eventual problem that needs
to be solved


Former-commit-id: 043a24b03b82e9b5f2549c446f732b31c68247fe