Use gen_block_mesh instead of Gmsh
Address #49

Use gen_block_mesh instead of Gmsh for generating the mesh. This
greatly reduces the complexity of the code. Also, Gmsh was not
generating a file format that SfePy could read.

The shape of the property arrays in get_material was incorrect and the
arrays were not contiguous. This has been fixed.

Calling predict on the model runs to completion without an error and
seems to converge. The results have not been checked yet though.


Former-commit-id: cb3c75dc739059e8bf2993fb6287e9015bb4fb58