Fix Boundary Conditions and resize_coeff
Fix issue #67

Fix the boundary conditions for the ElasticFEModel to have the
macroscopic strain value for any size microstructure. The previous
boundary conditions were incorrect and were creating an error with
re-sized influence coefficients.

Fix zero padding of the constant coefficients. For the constant valued
influence coefficients, the k0 value stays the same regardless of the
re-sizing.


Former-commit-id: 516d149051fcbefc275272d7924da21a78efb513