Show 2D model works while 3D model does not
Address issue #66

The files elasticFEModel.py and elasticFEModel3D.py currently print
out the displacement fields in 2D (ElasticFEModel) and in 3D
(ElasticFE3DModel).  In both models, the displacement fields should
run from 0.0 to 3.0 in the x direction and be zero in the other
directions. The ElasticFEModel does give this results while the
Elastic3Dmodel doesn't.


Former-commit-id: a810f54ce68313ae384a517ee23c06653ac6a824