Merge ElasticFEmodel and ElasticFE3DModel and Update API
Fix issue #71
Address issue #66

Merge ElasticFEModel and ElasticFE3DModel to work for both 3D and 2D
microstructures. Update ElasticFEModel API to take in material
properties as arguments when it is instantiated. and pass only the
microstructure.  Changed the name of the predict method to
get_response, and now takes the microstructure as an argument rather
than the microstructure property array. Update doc tests with new
API. Make all helper methods underscore methods. Remove dx as an
argument for ElasticFEModel. Update tests with new API.


Former-commit-id: 8091ca7833e4dd3885879bbb3037f22e1fbbf90b