Fix 3D periodic displacement BCs and int order
Address #74

Issues:

 * 3D periodic displacement boundary conditions work.

 * There was a bug previously with the x-z and x-y edge vertex
   positions for simple displacement test cases. The offset condition
   was not satisfied on the x-z and z-y edges on the right plane. The
   bug was caused by having two competing periodic boundary conditions
   in the x-direction.

 * Furthermore, the order=3 Integral order was wrong. It should be
   4. In general, it needs to be 2 * approx_order. This meant that the
   3D results up until this commit have probably been highly
   inaccurate.

Fixes:

This commit removed the x-direction periodicity from those edge nodes
on the y-plane and z-plane boundary conditions. Thus, the only
x-direction periodicity is the offset condition on the edge nodes.
This commit address the issue by introducing two periodic boundary
conditions for the z-plane and y-plane boundary conditions. One
boundary condition is for the y and z-direction and includes the x-z
and x-y edges. The other periodic boundary condition is for the
x-direction, but doesn't include the right most edge vertices. The
subdomain_func method of the ElasticFESimulation class now takes a
max_x argument, which enables the exclusion of the left most x-plane
vertices. There is a quite a lot of code duplication, which should be
address in a refactor due to having two similar methods for the
periodic BCs.

Other stuff:

 * The test_elastic_FE_simulation_2D was removed as it was testing
   arbitrary values. It has been replaced with a doctest that tests
   that all the periodic boundary conditions are satisfied.


Former-commit-id: 169fa2e0cc4a2634733177530a8e936cf6837fe7