Delete redundant notebooks
Address #95

Including notebooks that are not finalized or complete or are just for
testing and diagnostics is confusing to users looking through the
notebooks. All the notebooks that are not included in the
documentation are now deleted. The Polynomials notebook looks like it
was used to implement the Legendre basis functions, which is now
complete. It evidently isn't finished yet. If a developer wants any of
these notebooks they should branch from the commit before this one and
maintain that branch on their own site.


Former-commit-id: b52e9ea8372b8547cd00a8766965cd5820ee6018