Add draw_correlations
Address issue #135

Add draw_correlations method as a way to look at all of the
autocorrelations and cross-correlations for a given microstructure
function.

The functions still need to be changed to change the dimensions of the
figure to handle microstructures with a large number of correlations.
Additionally the functionality to select the correlations to view still
needs to be developed.


Former-commit-id: 3f3969902c22c75b0259bb0da12c2edc7741b644