Inherit BaseEstimator and Scale Data
Address issue #152

Inherit BaseEstimator from sklearn in order to use tools from sklearn
such as GrindSearchCV. Change name of class variables to match variables
in api to get functions _get_parameters and _set_parameters to work.

Scale data to by adding scale from sklearn. This step is redundant for
some of the dimensionality reduction techniques such as PCA, but is
required for others such as TruncateSVD.


Former-commit-id: 13c512a00d64abdd3ca19647f200f4cb6f60e3f6