Remove stats methods
address issue #186

Remove stats methods. By combining both the stats_fit and fit method
as well as the stats_predict and predict method, the
MKSHomogenizationModel can be used with either microstructures of
spatial correlations with `GridsearchCV` from `sklearn`. This will
allow for the model parameters to be found without recomputing the
spatial correlations every time.


Former-commit-id: cee70aee9d7acccd8f56a8baf8223762f0703e61