Mac Installation Guide
address #212

Including Mac OS X in the installation guide accounts for another large potential user base of PyMKS who may not have a Windows or Linux machine.

Even though the Mac OS X installation blurb is nearly identical to that of the already present installation guide, it shows that OS X is supported by PyMKS.

The only changes this should have is the inclusion of a short text paragraph in the INSTALLATION.md file.


Former-commit-id: e6cb221ab166914723ab61a70153783709980649