Update docs, syntax and allow n_states list
address #141

Update the doctest in GSH basis. Add __init__ method for Fourier basis
that gives meaningful initial values. Update syntax and remove
commented out sections of code from mks_localization.

Allow a list to be passed to n_states as well as an integer. This
change allows users to select the local states they want to use for
both homogenization and localization.


Former-commit-id: 771aabe425fbc41b8af9f3eecd12de31ec891280