Move FFT functions to bases.
address #232

Move FFT functions to bases. This change allows for fft functions to be
located in one locations. Depending on the whether the basis function
is complex or real fft or realfft can be used and reduce the amount
of memory used.

As a result of this change, the API for the stats functions are
different this change to the API removes the unnecessary creation
of the microstruture function variables. As a result this will also
save memory.


Former-commit-id: 79412bfb4e6fdf4069b04e3a02cfbe8630826784