Remove array resizing functionality from FFT functions
address #232

Remove array resizing functionality from FFT functions. This change
allows for the arrays to not be copied when performing the FFT
calculations. As a result of this change all of the zero padding is
done with the resize function that already existed in the Filter class.
This change might decrease the time taken to compute arrays slightly,
but will significantly improve the memory performance.


Former-commit-id: b880735837aa0fa54402bb40e24971f15f9d24d5