Move periodic_axes argument from fit transform and predict to init
address #228

Remove periodic_axes as a parameter passed to the fit, fit_transform,
transform and predict methods to a parameter passed to the init method.
This changes the API, but provides a clear function call for these
methods and decreases the chance that the correct values for periodic
axes will be selected.


Former-commit-id: 5041a855d4038d87c0a4ec033801240065615a04