Change size used for resizing data while using GridSearchCV
address #287

Change size used for resizing data while using GridSearchCV.
GridSearchCV requires that the inputs be 2 dimensions (n_samples,
n_features). This requires that the micorstructures need to be reshaped
after they have been passed back into the `MKSHomogenizationModel`.

This change ensures that the size used to resize the microstructures
is correct regardless of whether or not periodicity of the statistics.


Former-commit-id: dd07e5cdf47087e85799ad51dea59fb1f9f731f0